SqlTable improvments (you can SELECT * FROM <table> now). Started to implementet the session manager. It loads all the session from the database.

+++ src/webserver/UserDataBase.h
@@ -3,7 +3,7 @@
 #include <variant>
 #include <map>
 
-typedef std::variant<int, float, std::string> UserDataValue;
+typedef std::variant<int, double, std::string> UserDataValue;
 
 class UserDataBase{
